Consolidate coercion code in `datafusion_expr::type_coercion` and submodules (#3728)
* Move function coercion to its own module

* Move binary into type coercion

* fmt

* More updates

* consolidate some more

* Move aggregates
