+++ datafusion/physical-expr/src/expressions/binary.rs
@@ -76,7 +76,7 @@ use arrow::record_batch::RecordBatch;
 use crate::PhysicalExpr;
 use datafusion_common::ScalarValue;
 use datafusion_common::{DataFusionError, Result};
-use datafusion_expr::binary_rule::binary_operator_data_type;
+use datafusion_expr::type_coercion::binary::binary_operator_data_type;
 use datafusion_expr::{ColumnarValue, Operator};
 
 /// Binary expression
@@ -938,7 +938,7 @@ mod tests {
     use crate::expressions::{col, lit};
     use arrow::datatypes::{ArrowNumericType, Field, Int32Type, SchemaRef};
     use datafusion_common::Result;
-    use datafusion_expr::binary_rule::coerce_types;
+    use datafusion_expr::type_coercion::binary::coerce_types;
 
     // Create a binary expression without coercion. Used here when we do not want to coerce the expressions
     // to valid types. Usage can result in an execution (after plan) error.
